@@ -42,8 +42,7 @@ class DataColors {
         return (
             `#${(0x1000000 + ((R < 255 ? (R < 1 ? 0 : R) : 255) * 0x10000) +
                     ((G < 255 ? (G < 1 ? 0 : G) : 255) * 0x100) +
-                    (B < 255 ? (B < 1 ? 0 : B) : 255))
-                    .toString(16).slice(1)}`
+                    (B < 255 ? (B < 1 ? 0 : B) : 255)).toString(16).slice(1)}`
         );
     }
 
