@@ -58,7 +58,6 @@ export class FileTable extends React.Component {
 
         // Initialize component state.
         this.state = {
-            maxWidth: 'auto', // Width of widest table
             collapsed: { // Keeps track of which tables are collapsed
                 raw: false,
                 rawArray: false,
@@ -81,15 +80,17 @@ export class FileTable extends React.Component {
     }
 
     fileClick(file) {
-        const node = {
-            '@type': ['File'],
-            metadata: {
-                ref: file,
-            },
-            schemas: this.props.schemas,
-        };
-        this.props.setInfoNodeId(node);
-        this.props.setInfoNodeVisible(true);
+        if (this.props.setInfoNodeId && this.props.setInfoNodeVisible) {
+            const node = {
+                '@type': ['File'],
+                metadata: {
+                    ref: file,
+                },
+                schemas: this.props.schemas,
+            };
+            this.props.setInfoNodeId(node);
+            this.props.setInfoNodeVisible(true);
+        }
     }
 
     handleCollapse(table) {
@@ -194,7 +195,8 @@ export class FileTable extends React.Component {
                         <SortTable
                             title={
                                 <CollapsingTitle
-                                    title="Processed data" collapsed={this.state.collapsed.proc}
+                                    title="Processed data"
+                                    collapsed={this.state.collapsed.proc}
                                     handleCollapse={this.handleCollapseProc}
                                 />
                             }
@@ -260,6 +262,22 @@ FileTable.propTypes = {
     adminUser: PropTypes.bool, // True if user is an admin user
     schemas: PropTypes.object, // Object from /profiles/ containing all schemas
     noDefaultClasses: PropTypes.bool, // True to strip SortTable panel of default CSS classes
+};
+
+FileTable.defaultProps = {
+    graphedFiles: null,
+    filePanelHeader: null,
+    encodevers: '',
+    selectedFilterValue: 'default',
+    filterOptions: [],
+    anisogenic: false,
+    showFileCount: false,
+    setInfoNodeId: null,
+    setInfoNodeVisible: null,
+    session: null,
+    adminUser: false,
+    schemas: null,
+    noDefaultClasses: false,
 };
 
 FileTable.contextTypes = {
@@ -432,21 +450,15 @@ class RawSequencingTable extends React.Component {
         // Initialize React state variables.
         this.state = {
             collapsed: false, // Collapsed/uncollapsed state of table
-            restrictedTip: '', // UUID of file with tooltip showing
         };
 
         // Bind `this` to non-React methods.
         this.handleCollapse = this.handleCollapse.bind(this);
-        this.hoverDL = this.hoverDL.bind(this);
     }
 
     handleCollapse() {
         // Handle a click on a collapse button by toggling the corresponding tableCollapse state var
         this.setState({ collapsed: !this.state.collapsed });
-    }
-
-    hoverDL(hovering, fileUuid) {
-        this.setState({ restrictedTip: hovering ? fileUuid : '' });
     }
 
     render() {
@@ -650,7 +662,11 @@ class RawSequencingTable extends React.Component {
 
 RawSequencingTable.propTypes = {
     files: PropTypes.array, // Raw files to display
-    meta: PropTypes.object, // Extra metadata in the same format passed to SortTable
+    meta: PropTypes.object.isRequired, // Extra metadata in the same format passed to SortTable
+};
+
+RawSequencingTable.defaultProps = {
+    files: null,
 };
 
 
@@ -661,21 +677,15 @@ class RawFileTable extends React.Component {
         // Initialize React state variables.
         this.state = {
             collapsed: false, // Collapsed/uncollapsed state of table
-            restrictedTip: '', // UUID of file with tooltip showing
         };
 
         // Bind `this` to non-React methods.
         this.handleCollapse = this.handleCollapse.bind(this);
-        this.hoverDL = this.hoverDL.bind(this);
     }
 
     handleCollapse() {
         // Handle a click on a collapse button by toggling the corresponding tableCollapse state var
         this.setState({ collapsed: !this.state.collapsed });
-    }
-
-    hoverDL(hovering, fileUuid) {
-        this.setState({ restrictedTip: hovering ? fileUuid : '' });
     }
 
     render() {
@@ -824,7 +834,11 @@ class RawFileTable extends React.Component {
 
 RawFileTable.propTypes = {
     files: PropTypes.array, // Raw sequencing files to display
-    meta: PropTypes.object, // Extra metadata in the same format passed to SortTable
+    meta: PropTypes.object.isRequired, // Extra metadata in the same format passed to SortTable
+};
+
+RawFileTable.defaultProps = {
+    files: null,
 };
 
 
@@ -844,11 +858,16 @@ DatasetFiles.propTypes = {
     items: PropTypes.array, // Array of files retrieved
 };
 
+DatasetFiles.defaultProps = {
+    items: null,
+};
+
 
 // File display widget, showing a facet list, a table, and a graph (and maybe a BioDalliance).
 // This component only triggers the data retrieval, which is done with a search for files associated
 // with the given experiment (in this.props.context). An odd thing is we specify query-string parameters
 // to the experiment URL, but they apply to the file search -- not the experiment itself.
+/* eslint-disable react/prefer-stateless-function */
 export class FileGallery extends React.Component {
     render() {
         const { context, encodevers, anisogenic, hideGraph, altFilterDefault } = this.props;
@@ -862,13 +881,21 @@ export class FileGallery extends React.Component {
         );
     }
 }
+/* eslint-enable react/prefer-stateless-function */
 
 FileGallery.propTypes = {
-    context: PropTypes.object, // Dataset object whose files we're rendering
+    context: PropTypes.object.isRequired, // Dataset object whose files we're rendering
     encodevers: PropTypes.string, // ENCODE version number
     anisogenic: PropTypes.bool, // True if anisogenic experiment
     hideGraph: PropTypes.bool, // T to hide graph display
     altFilterDefault: PropTypes.bool, // T to default to All Assemblies and Annotations
+};
+
+FileGallery.defaultProps = {
+    encodevers: '',
+    anisogenic: false,
+    hideGraph: false,
+    altFilterDefault: false,
 };
 
 FileGallery.contextTypes = {
@@ -1643,11 +1670,8 @@ class FileGalleryRendererComponent extends React.Component {
         // Initialize React state variables.
         this.state = {
             selectedFilterValue: 'default', // <select> value of selected filter
-            meta: null, // @id of node whose info panel is open
-            infoModalOpen: false, // True if info modal is open
             relatedFiles: [],
             inclusionOn: adminUser, // True to exclude files with certain statuses
-            contributingFiles: [], // Cache for contributing files retrieved from the DB
         };
 
         // Bind `this` to non-React methods.
@@ -1871,13 +1895,20 @@ FileGalleryRendererComponent.inclusionStatuses = [
 ];
 
 FileGalleryRendererComponent.propTypes = {
-    context: PropTypes.object, // Dataset whose files we're rendering
+    context: PropTypes.object.isRequired, // Dataset whose files we're rendering
     data: PropTypes.object, // File data retrieved from search request
     schemas: PropTypes.object, // Schemas for the entire system; used for QC property titles
     hideGraph: PropTypes.bool, // T to hide graph display
     altFilterDefault: PropTypes.bool, // T to default to All Assemblies and Annotations
-    auditIndicators: PropTypes.func, // Inherited from auditDecor HOC
-    auditDetail: PropTypes.func, // Inherited from auditDecor HOC
+    auditIndicators: PropTypes.func.isRequired, // Inherited from auditDecor HOC
+    auditDetail: PropTypes.func.isRequired, // Inherited from auditDecor HOC
+};
+
+FileGalleryRendererComponent.defaultProps = {
+    data: null,
+    schemas: null,
+    hideGraph: false,
+    altFilterDefault: false,
 };
 
 FileGalleryRendererComponent.contextTypes = {
@@ -1905,6 +1936,10 @@ CollapsingTitle.propTypes = {
     collapsed: PropTypes.bool, // T if the panel this is over has been collapsed
 };
 
+CollapsingTitle.defaultProps = {
+    collapsed: false,
+};
+
 
 // Display a filtering <select>. `filterOptions` is an array of objects with two properties:
 // `assembly` and `annotation`. Both are strings that get concatenated to form each menu item. The
@@ -1917,7 +1952,7 @@ const FilterMenu = (props) => {
             <option value="default">All Assemblies and Annotations</option>
             <option disabled="disabled" />
             {filterOptions.map((option, i) =>
-                <option key={`${option.assembly}${option.annotation}`} value={i}>{`${option.assembly + (option.annotation ? ` ${option.annotation}` : '')}`}</option>,
+                <option key={`${option.assembly}${option.annotation}`} value={i}>{`${option.assembly + (option.annotation ? ` ${option.annotation}` : '')}`}</option>
             )}
         </select>
     );
@@ -1927,6 +1962,10 @@ FilterMenu.propTypes = {
     selectedFilterValue: PropTypes.string, // Currently selected filter
     filterOptions: PropTypes.array.isRequired, // Contents of the filtering menu
     handleFilterChange: PropTypes.func.isRequired, // Call when a filtering option changes
+};
+
+FilterMenu.defaultProps = {
+    selectedFilterValue: 'default',
 };
 
 
@@ -2111,7 +2150,7 @@ const FileDetailView = function FileDetailView(node, qcClick, auditIndicators, a
                             <dt>File quality metrics</dt>
                             <dd className="file-qc-buttons">
                                 {selectedFile.quality_metrics.map(qc =>
-                                    <FileQCButton key={qc['@id']} qc={qc} file={selectedFile} schemas={node.schemas} handleClick={qcClick} />,
+                                    <FileQCButton key={qc['@id']} qc={qc} file={selectedFile} schemas={node.schemas} handleClick={qcClick} />
                                 )}
                             </dd>
                         </div>
